add: auto save custom hook
